fix: Update 'Price' property in 'Request.cs'
The 'Request.cs' file in the 'Flipster.Modules.Catalog/Endpoints/Adverts/Change' directory has been modified to include a change in the 'Price' property. Previously, the 'Price' property was declared and assigned a value within the record declaration, but it has now been moved outside of the record declaration and is declared as a public property. Additionally, a default value of 0.00m has been assigned to the 'Price' property. This change allows for external access and modification of the 'Price' property in the 'Request' record.