feat: Add delete endpoint for adverts
This change adds a new delete endpoint for adverts in the AdvertsEndpoints class. The delete endpoint is only accessible to authenticated users as it requires a valid JWT token. The endpoint allows users to delete an advert by providing the advert's ID in the route parameter. The delete operation is only allowed if the current user is the owner of the advert or has an admin role. If the advert with the given ID is not found or the user is not authorized to delete the advert, appropriate error messages are thrown. After successfully deleting the advert, an empty response is returned.

This change enhances the functionality of the AdvertsEndpoints class by introducing the ability to delete existing adverts through the API.