feat: Add FavoriteRepository
This commit adds the `FavoriteRepository` implementation in the `Flipster.Modules.Users.Infrastructure` project. The repository includes methods to add, remove, and update `Favorite` entities. It also provides methods to get a favorite by its ID, get all favorites, and get favorites by user ID and advert ID.

The `FavoriteRepository` is now registered as a transient service in the `DependencyInjection` class in the same project.

The `User` entity in the `Flipster.Modules.Users.Domain` project has been updated to include a collection of `Favorite` entities. The `UserConfiguration` class in the `Flipster.Modules.Users.Infrastructure` project has been modified to include a relationship between `User` and `Favorite`.

The `UsersModuleDbContext` class in the `Flipster.Modules.Users.Infrastructure` project now includes a `DbSet` for `Favorite` entities.

---

*Note: This commit also includes minor changes to file imports and a removal of unused code in the `Flipster.Modules.Users` project.*